Working framework!  In theory, all required build info is in package.json.
Run browser tests with Jasmine standalone test/ScriptRunner.html, and node tests with the 'npm
test' script.

TODO: write all the unit tests, build up commune to where it was at FDG,
duplicate it over in node.js

Potential Holes:
--getting event handling by working the DOM.  Probably not doable in Node
contexts, but that's probably ok? Considering that, right now, events
are only fired off by WebArtStores
--web art bots, right now, can only suscribe to one art store
--lots of isomorphic mongoose hacks to keep similar schema for both node
and browser
--on that note, the schema for Tags means that you can't access them
like a dictionary, but they really want to be a dictionary. Move to a
mixed type rather than an array of pairs of strings?