tweaked directory structure(completely cleared out non-v3 things, moved the __v3 folder to main directory removed the __v3 folder), added a testing framework and some unit tests.
Current Large Issues:
I forgot that JavaScript does a horrid job with multiple
inheritance, and manually coded a solution.  The correct action might be to try
and define a stateful mixin for publishing rather than a full object
with a constructor?

Fudimental problem is that WebArtist and VisualArtist implement different parts of the Artist interface, and having a class inherit
from both is tricky at best.

The dependancy graph for node to get pixels from an SVG string is
too large.  Figure out an easy way to integrate phantomjs for it, or
some other solution (right now requires jsdom, node-canvas, imagemagick
and graphicsmagick)

def need to do a UI/UX update, a lot more testing, add in
evaluators,start creating node artists and artstores